versoview::compositor
Struct IOCompositor

pub struct IOCompositor {
    pub current_window: WindowId,
    pub ready_to_present: bool,
    pub shutdown_state: ShutdownState,
    pub constellation_chan: Sender<ConstellationMsg>,
    pub webrender_api: RenderApi,
    pub rendering_context: RenderingContext,
    pub is_animating: bool,
    /* private fields */
}

The Verso compositor contains a GL rendering context with a WebRender instance. The compositor will communicate with Servo using messages from the Constellation, then composite the WebRender frames and present the surface to the window.

Fields
current_window: WindowId

The current window that Compositor is handling.

ready_to_present: bool

check if the surface is ready to present.

shutdown_state: ShutdownState

Tracks whether we are in the process of shutting down, or have shut down and should close the compositor.

constellation_chan: Sender<ConstellationMsg>

The channel on which messages can be sent to the constellation.

webrender_api: RenderApi

The webrender interface, if enabled.

rendering_context: RenderingContext

The glutin instance that webrender targets

is_animating: bool

Whether the application is currently animating. Typically, when animations are active, the window will want to avoid blocking on UI events, and just run the event loop at the vsync interval.

Implementations

impl IOCompositor

pub fn new( current_window: WindowId, viewport: DeviceIntSize, scale_factor: Scale<f32, DeviceIndependentPixel, DevicePixel>, state: InitialCompositorState, exit_after_load: bool, convert_mouse_to_touch: bool, ) -> Self

Create a new compositor.

pub fn deinit(&mut self)

Consume compositor itself and deinit webrender.

pub fn maybe_start_shutting_down(&mut self)

Tell compositor to start shutting down.

pub fn send_root_pipeline_display_list(&mut self, window: &Window)

Set the root pipeline for our WebRender scene to a display list that consists of an iframe for each visible top-level browsing context, applying a transformation on the root for pinch zoom, page zoom, and HiDPI scaling.

pub fn on_resize_webview_event( &mut self, webview_id: TopLevelBrowsingContextId, rect: DeviceIntRect, )

Notify compositor the provided webview is resized. The compositor will tell constellation and update the display list.

pub fn swap_current_window(&mut self, window: &mut Window)

Change the current window of the compositor should display.

pub fn resize(&mut self, size: Size2D<i32, DevicePixel>, window: &mut Window)

Resize the rendering context and all web views.

pub fn on_resize_window_event( &mut self, new_viewport: DeviceIntSize, window: &Window, )

Handle the window resize event.

pub fn on_scale_factor_event( &mut self, scale_factor: f32, window: &Window, ) -> bool

Handle the window scale factor event and return a boolean to tell embedder if it should further handle the scale factor event.

pub fn on_input_event(&mut self, event: InputEvent)

Handle the input event in the window.

pub fn on_touch_event(&mut self, event: TouchEvent)

Handle touch event.

pub fn on_scroll_event( &mut self, scroll_location: ScrollLocation, cursor: DeviceIntPoint, action: TouchEventAction, )

Handle scroll event.

pub fn on_zoom_reset_window_event(&mut self, window: &Window)

Handle zoom reset event

pub fn on_zoom_window_event(&mut self, magnification: f32, window: &Window)

Handle zoom event in the window

pub fn on_pinch_zoom_window_event(&mut self, magnification: f32)

Simulate a pinch zoom

pub fn composite(&mut self, window: &Window)

Composite to the given target if any, or the current target otherwise.

pub fn receive_messages( &mut self, windows: &mut HashMap<WindowId, (Window, DocumentId)>, ) -> bool

Receive and handle compositor messages.

pub fn perform_updates( &mut self, windows: &mut HashMap<WindowId, (Window, DocumentId)>, ) -> bool

Perform composition and related actions.

pub fn toggle_webrender_debug(&mut self, option: WebRenderDebugOption)

Update debug option of the webrender.