versoview::tab
Struct TabManager

pub struct TabManager { /* private fields */ }

Tab manager to handle multiple tab in a window.

Implementations

impl TabManager

pub fn new() -> Self

Create a new tab manager.

pub fn count(&self) -> usize

Get tab count.

pub fn current_tab_id(&self) -> Option<WebViewId>

Get current actvie tab id.

pub fn current_tab(&self) -> Option<&Tab>

Get current active tab.

pub fn tab_ids(&self) -> Vec<WebViewId>

Get all tab id.

pub fn activate_tab(&mut self, tab_id: WebViewId) -> Option<&Tab>

Activate the tab by tab id.

pub fn tab(&self, id: WebViewId) -> Option<&Tab>

Get tab by tab id.

pub fn append_tab(&mut self, webview: WebView, active: bool)

Append a tab.

pub fn close_tab(&mut self, id: WebViewId) -> Result<Tab, TabManagerErr>

Close a tab.

pub fn set_size( &mut self, tab_id: WebViewId, rect: DeviceIntRect, ) -> (Option<WebViewId>, Option<WebViewId>)

Set tab size. Will also set prompt dialog size if it exists.

Returns the tab and prompt WebViewId if they exist.

pub fn history(&self, tab_id: WebViewId) -> Option<&TabHistory>

Get tab history.

pub fn set_history( &mut self, tab_id: WebViewId, list: Vec<ServoUrl>, current_idx: usize, )

Set tab history.

pub fn prompt_by_tab_id(&self, tab_id: WebViewId) -> Option<&PromptDialog>

Get prompt dialog by tab id.

pub fn prompt_by_prompt_id(&self, prompt_id: WebViewId) -> Option<&PromptDialog>

Get prompt dialog by tab id.

pub fn current_prompt(&self) -> Option<&PromptDialog>

Get current tabw prompt dialog.

pub fn set_prompt(&mut self, tab_id: WebViewId, prompt: PromptDialog)

Set tab prompt dialog.

pub fn remove_prompt_by_tab_id( &mut self, tab_id: WebViewId, ) -> Option<PromptDialog>

Remove prompt by tab webview ID.

pub fn remove_prompt_by_prompt_id( &mut self, prompt_id: WebViewId, ) -> Option<PromptDialog>

Remove prompt by prompt webview ID.

pub fn has_prompt(&self, prompt_id: WebViewId) -> bool

Check if there is a prompt dialog by prompt webview ID.