versoview::window
Struct Window

pub struct Window { /* private fields */ }

A Verso window is a Winit window containing several web views.

Implementations

impl Window

pub fn handle_servo_messages_with_webview( &mut self, webview_id: WebViewId, message: EmbedderMsg, sender: &Sender<ConstellationMsg>, clipboard: Option<&mut Clipboard>, compositor: &mut IOCompositor, )

Handle servo messages with corresponding web view ID.

pub fn handle_servo_messages_with_panel( &mut self, panel_id: WebViewId, message: EmbedderMsg, sender: &Sender<ConstellationMsg>, clipboard: Option<&mut Clipboard>, compositor: &mut IOCompositor, ) -> bool

Handle servo messages with main panel. Return true it requests a new window.

pub fn handle_servo_messages_with_context_menu( &mut self, webview_id: WebViewId, message: EmbedderMsg, sender: &Sender<ConstellationMsg>, _clipboard: Option<&mut Clipboard>, _compositor: &mut IOCompositor, ) -> bool

Handle servo messages with main panel. Return true it requests a new window.

pub fn handle_servo_messages_with_prompt( &mut self, webview_id: WebViewId, message: EmbedderMsg, _sender: &Sender<ConstellationMsg>, _clipboard: Option<&mut Clipboard>, _compositor: &mut IOCompositor, ) -> bool

Handle servo messages with prompt. Return true it requests a new window.

impl Window

pub fn new( evl: &ActiveEventLoop, window_attributes: WindowAttributes, ) -> (Self, RenderingContext)

Create a Verso window from Winit window and return the rendering context.

pub fn new_with_compositor( evl: &ActiveEventLoop, compositor: &mut IOCompositor, ) -> Self

Create a Verso window with the rendering context.

pub fn get_content_size( &self, size: DeviceIntRect, include_tab: bool, ) -> DeviceIntRect

Get the content area size for the webview to draw on

pub fn create_panel( &mut self, constellation_sender: &Sender<ConstellationMsg>, initial_url: Option<Url>, )

Send the constellation message to start Panel UI

pub fn create_tab( &mut self, constellation_sender: &Sender<ConstellationMsg>, initial_url: ServoUrl, )

Create a new webview and send the constellation message to load the initial URL

pub fn close_tab(&mut self, compositor: &mut IOCompositor, tab_id: WebViewId)

Close a tab

pub fn activate_tab( &mut self, compositor: &mut IOCompositor, tab_id: WebViewId, show_tab: bool, )

Activate a tab

pub fn handle_winit_window_event( &mut self, sender: &Sender<ConstellationMsg>, compositor: &mut IOCompositor, event: &WindowEvent, )

Handle Winit window event and return a boolean to indicate if the compositor should repaint immediately.

pub fn handle_servo_message( &mut self, webview_id: WebViewId, message: EmbedderMsg, sender: &Sender<ConstellationMsg>, clipboard: Option<&mut Clipboard>, compositor: &mut IOCompositor, ) -> bool

Handle servo messages. Return true if it requests a new window

pub fn request_redraw(&self)

Queues a Winit WindowEvent::RedrawRequested event to be emitted that aligns with the windowing system drawing loop.

pub fn size(&self) -> DeviceIntSize

Size of the window that’s used by webrender.

pub fn id(&self) -> WindowId

Get Winit window ID of the window.

pub fn scale_factor(&self) -> f64

Scale factor of the window. This is also known as HIDPI.

pub fn has_webview(&self, id: WebViewId) -> bool

Check if the window has such webview.

pub fn remove_webview( &mut self, id: WebViewId, compositor: &mut IOCompositor, ) -> (Option<WebView>, bool)

Remove the webview in this window by provided webview ID. If provided ID is the panel, it will shut down the compositor and then close whole application.

pub fn painting_order(&self) -> Vec<&WebView>

Get the painting order of this window.

pub fn set_cursor_icon(&self, cursor: Cursor)

Set cursor icon of the window.